pyproject.toml
[project]
name = "memoery-for-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-genai>=1.62.0",
    "pymongo>=4.6.0",
    "psycopg2-binary>=2.9.9",
    "openai>=2.17.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "huggingface-hub>=1.4.1",
]
